image uploads are done for both products and carousel

+++ backend/controllers/carouselController.js
@@ -15,39 +15,46 @@ const getImages = asyncHandler(async (req, res) => {
     if (!data) {
         res.status(500).json({ message: 'server or DB error please try again' })
         return
+    } else {
+        res.status(200).json(data)
     }
-    res.status(200).json(data)
 })
 
 // @desc    Add new image
 // @route   POST /api/carousel
 // @access  private
 const addImage = asyncHandler(async (req, res) => {
-    const {
-        imageURL,
-        productURL,
-        isActive,
-    } = req.body
+    if (req.user.type === 'Admin' && req.user.status === 'Active') {
+        const {
+            imageURL,
+            productURL,
+            isActive,
+        } = req.body
 
-    const newImage = {
-        imageURL,
-        productURL,
-        isActive,
-    }
+        const newImage = {
+            imageURL,
+            productURL: productURL ? productURL : '',
+            isActive,
+        }
 
-    // check if image exists
-    const exists = await Carousel.findOne({ imageURL: newImage.imageURL })
-    if (exists) {
-        res.status(400).json({ message: 'already exists' })
-    }
-
-    // create the product
-    try {
-        const data = await Carousel.create(newImage)
-        res.status(201).json(data)
-    } catch (err) {
+        // check if image exists
+        const exists = await Carousel.findOne({ imageURL: newImage.imageURL })
+        if (exists) {
+            res.status(400)
+            throw new Error('image already exists')
+        } else {
+            // create the image
+            try {
+                const data = await Carousel.create(newImage)
+                res.status(201).json(data)
+            } catch (err) {
+                res.status(401)
+                throw new Error('unknowen server or DB error')
+            }
+        }
+    } else {
         res.status(401)
-        res.json(err)
+        throw new Error(`Unauthorized, no privilges`)
     }
 })
 
@@ -55,19 +62,24 @@ const addImage = asyncHandler(async (req, res) => {
 // @route   DELETE /api/carousel/:id
 // @access  private
 const deleteImage = asyncHandler(async (req, res) => {
-    const id = req.params.id
+    if (req.user.type === 'Admin' && req.user.status === 'Active') {
+        const id = req.params.id
 
-    // Check for product
-    const doc = await Carousel.findById(id)
+        // Check for image
+        const doc = await Carousel.findById(id)
 
-    if (doc) {
-        await Carousel.deleteOne({ _id: id })
-        res.status(201).json({
-            id: doc.id
-        })
+        if (doc) {
+            await Carousel.deleteOne({ _id: id })
+            res.status(200).json({
+                id: doc.id
+            })
+        } else {
+            res.status(400)
+            throw new Error('Invalid image id')
+        }
     } else {
-        res.status(400)
-        throw new Error('Invalid image id')
+        res.status(401)
+        throw new Error(`Unauthorized, no privilges`)
     }
 })
 
@@ -75,30 +87,35 @@ const deleteImage = asyncHandler(async (req, res) => {
 // @route   PUT /api/carousel/:id
 // @access  private
 const editImage = asyncHandler(async (req, res) => {
-    const {
-        imageURL,
-        productURL,
-        isActive,
-    } = req.body
-    const id = req.params.id
-
-    // Check for image
-    const doc = await Carousel.findById(id)
-
-    if (doc) {
-        const data = await Carousel.findOneAndUpdate({ _id: id }, {
+    if (req.user.type === 'Admin' && req.user.status === 'Active') {
+        const {
             imageURL,
             productURL,
             isActive,
-        }, {
-            new: true
-        })
-        res.status(200).json({
-            updated: data
-        })
+        } = req.body
+        const id = req.params.id
+
+        // Check for image
+        const doc = await Carousel.findById(id)
+
+        if (doc) {
+            const data = await Carousel.findOneAndUpdate({ _id: id }, {
+                imageURL,
+                productURL,
+                isActive,
+            }, {
+                new: true
+            })
+            res.status(200).json({
+                updated: data
+            })
+        } else {
+            res.status(400)
+            throw new Error('Invalid image id')
+        }
     } else {
-        res.status(400)
-        throw new Error('Invalid image id')
+        res.status(401)
+        throw new Error(`Unauthorized, no privilges`)
     }
 })
